ISSUE #2901 further fixes
