+++ backend/tests/v5/unit/middleware/dataConverter/inputs/teamspaces/projects/models/containers.test.js
@@ -30,6 +30,9 @@ const Responder = require(`${src}/utils/responder`);
 jest.mock('../../../../../../../../../src/v5/models/modelSettings');
 const ModelSettings = require(`${src}/models/modelSettings`);
 
+jest.mock('../../../../../../../../../src/v5/models/revisions');
+const RevisionsModel = require(`${src}/models/revisions`);
+
 const Containers = require(`${src}/middleware/dataConverter/inputs/teamspaces/projects/models/containers`);
 const { templates } = require(`${src}/utils/responseCodes`);
 
@@ -64,15 +67,20 @@ const testCanDeleteContainer = () => {
 	});
 };
 
-const createRequestWithFile = (teamspace, { tag, desc, importAnim }, unsupportedFile = false, noFile = false) => {
+const createRequestWithFile = (teamspace, { tag, desc, importAnim, timezone },
+	unsupportedFile = false, noFile = false, emptyFile = false) => {
 	const form = new FormData();
 	if (!noFile) {
+		let filePath = unsupportedFile ? 'dummy.png' : 'dummy.obj';
+		filePath = emptyFile ? 'empty.ifc' : filePath;
+
 		form.append('file',
-			fs.createReadStream(path.join(modelFolder, unsupportedFile ? 'dummy.png' : 'dummy.obj')));
+			fs.createReadStream(path.join(modelFolder, filePath)));
 	}
 	if (tag) form.append('tag', tag);
 	if (desc) form.append('desc', desc);
 	if (importAnim) form.append('importAnim', importAnim);
+	if (timezone) form.append('timezone', timezone);
 
 	const req = new MockExpressRequest({
 		method: 'POST',
@@ -87,21 +95,27 @@ const createRequestWithFile = (teamspace, { tag, desc, importAnim }, unsupported
 };
 
 Quota.sufficientQuota.mockImplementation((ts) => (ts === 'noQuota' ? Promise.reject(templates.quotaLimitExceeded) : Promise.resolve()));
+RevisionsModel.isTagUnique.mockImplementation((teamspace, model, tag) => tag !== 'duplicate');
 
 const testValidateNewRevisionData = () => {
-	const standardBody = { tag: '123', description: 'this is a model', importAnim: false };
+	const standardBody = { tag: '123', description: 'this is a model', importAnimations: false, timezone: 'Europe/Berlin' };
 	describe.each([
 		['Request with valid data', 'ts', standardBody],
-		['Request with unsupported model file', 'ts', standardBody, true, false, templates.unsupportedFileFormat],
-		['Request with insufficient quota', 'noQuota', standardBody, false, false, templates.quotaLimitExceeded],
-		['Request with no body should fail', 'ts', {}, false, false, templates.invalidArguments],
-		['Request with just tag should pass', 'ts', { tag: 'dkf_j-d' }, false, false],
-		['Request with wrong tag type should fail', 'ts', { tag: false }, false, false, templates.invalidArguments],
-		['Request with tag that is not alphanumeric should fail', 'ts', { tag: '1%2%3' }, false, false, templates.invalidArguments],
-		['Request with no file should fail', 'ts', { tag: 'drflgdf' }, false, true, templates.invalidArguments],
-	])('Check new revision data', (desc, ts, bodyContent, badFile, noFile, error) => {
+		['Request with unsupported model file', 'ts', standardBody, true, false, false, templates.unsupportedFileFormat],
+		['Request with insufficient quota', 'noQuota', standardBody, false, false, false, templates.quotaLimitExceeded],
+		['Request with no body should fail', 'ts', {}, false, false, false, templates.invalidArguments],
+		['Request with just tag should pass', 'ts', { tag: 'dkf_j-d' }, false, false, false],
+		['Request with wrong tag type should fail', 'ts', { tag: false }, false, false, false, templates.invalidArguments],
+		['Request with tag that is not alphanumeric should fail', 'ts', { tag: '1%2%3' }, false, false, false, templates.invalidArguments],
+		['Request with no file should fail', 'ts', { tag: 'drflgdf' }, false, true, false, templates.invalidArguments],
+		['Request with an empty file should fail', 'ts', { tag: 'drflgdf' }, false, false, true, templates.invalidArguments],
+		['Request with duplicate tag should fail', 'ts', { tag: 'duplicate' }, false, false, false, templates.invalidArguments],
+		['Request with invalid timezone should fail', 'ts', { tag: 'drflgdf', timezone: 'abc' }, false, false, false, templates.invalidArguments],
+		['Request with invalid type timezone should fail', 'ts', { tag: 'drflgdf', timezone: 123 }, false, false, false, templates.invalidArguments],
+		['Request with null timezone should pass', 'ts', { tag: 'drflgdf', timezone: null }],
+	])('Check new revision data', (desc, ts, bodyContent, badFile, noFile, emptyFile, error) => {
 		test(`${desc} should ${error ? `fail with ${error.code}` : ' succeed and next() should be called'}`, async () => {
-			const req = createRequestWithFile(ts, bodyContent, badFile, noFile);
+			const req = createRequestWithFile(ts, bodyContent, badFile, noFile, emptyFile);
 			const mockCB = jest.fn(() => {});
 			await Containers.validateNewRevisionData(req, {}, mockCB);
 			if (error) {
